services/yt-api/requirements.txt
flask
Flask-SQLAlchemy
psycopg2-binary
celery
redis
python-dotenv
requests
python-dateutil
Flask-Caching
elasticsearch==7.9.1